automotive-service-system/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg[binary]
alembic
pydantic>=2.0
python-dotenv
